requirements.txt
# Core Bot Dependencies
python-telegram-bot==22.5
APScheduler==3.10.4
python-dotenv==1.0.1
SQLAlchemy==2.0.41
aiosqlite==0.22.1
cryptography==46.0.2
psutil==5.9.8
httpx==0.28.1
beautifulsoup4==4.12.3
gspread==6.2.1
google-auth==2.41.0
google-auth-oauthlib==1.2.3
pytz==2023.3
tzdata==2025.2
greenlet==3.2.3
typing_extensions==4.12.2
six==1.17.0
cffi==2.0.0
sniffio==1.3.1
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
requests==2.32.4
requests-oauthlib==2.0.0
oauthlib==3.3.1
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
cachetools==6.2.4
soupsieve==2.6
multidict==6.6.2
yarl==1.20.1
frozenlist==1.7.0
aiosignal==1.3.2
attrs==25.3.0
propcache==0.3.2
aiohappyeyeballs==2.6.1
asyncio==3.4.3
aiohttp==3.10.5
asyncpg==0.29.0
psycopg2-binary==2.9.9
